Extract withdraw presenter logic into module
Make withdrawn logic shareable between formats
